requirements.txt
# CLI Framework
click==8.1.7

# HTTP Client
requests==2.32.4
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Configuration
PyYAML==6.0.2

# Platform Support
platformdirs==4.4.0